upgrade tokenizers
